Tweaks and code refactoring
